optirutas-ia-backend/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
numpy
pandas
matplotlib
azure-storage-blob